Move react-table to its own component
